+++ packages/common/src/schemas/index.ts
@@ -4,6 +4,11 @@ export {
   defaultInputContextSchema,
   type DefaultContextSchema,
 } from './documents/default-context-schema';
+export { DocumentInsertSchema, DocumentsSchema } from './documents/schemas/documents-schema';
+export {
+  WorkflowRuntimeConfigSchema,
+  type TWorkflowRuntimeConfig,
+} from './documents/workflow/config-schema';
 export { getGhanaDocuments } from './documents/workflow/documents/schemas/GH';
 export {
   findDocumentSchemaByTypeAndCategory,
@@ -12,10 +17,5 @@ export {
   getDocumentsByCountry,
 } from './documents/workflow/documents/schemas/index';
 export { type TAvailableDocuments, type TDocument } from './documents/workflow/documents/types';
-export { WorkflowDefinitionConfigThemeSchema } from './workflow/workflow-config-theme';
 export * from './workflow/end-user.schema';
-export { DocumentsSchema, DocumentInsertSchema } from './documents/schemas/documents-schema';
-export {
-  WorkflowRuntimeConfigSchema,
-  type TWorkflowRuntimeConfig,
-} from './documents/workflow/config-schema';
+export { WorkflowDefinitionConfigThemeSchema } from './workflow/workflow-config-theme';