+++ packages/common/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/common
 
+## 0.9.46
+
+### Patch Changes
+
+- Bump
+
+## 0.9.45
+
+### Patch Changes
+
+- bump
+
+## 0.9.44
+
+### Patch Changes
+
+- Added collection flow manager & updated schema
+
 ## 0.9.43
 
 ### Patch Changes